Scope lib compile_error behind crypto features
